python-ms/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg[binary]
langchain>=0.0.300
openai
python-dateutil
httpx
